+++ bin/update.sh
@@ -28,7 +28,7 @@ max_capacity=${MAX}
 remaining_capacity=${MAX}
 
 group_start "Updating tools"
-debug "starting update capacity: ${remaining_capacity}"
+debug "starting with update capacity: ${remaining_capacity}"
 
 while read -r line; do
 	case "${line}" in
@@ -39,19 +39,21 @@ while read -r line; do
 		debug "skipping comment ('${line}')"
 		;;
 	*)
+		debug "processing line ('${line}')"
+
 		tool="$(echo "${line}" | awk '{print $1}')"
 		info "evaluating ${tool}..."
 
 		current_version="$(echo "${line}" | awk '{print $2}')"
 		latest_version="$(asdf latest "${tool}")"
-		debug "${tool} current version: ${current_version}"
-		debug "${tool} latest version: ${latest_version}"
+		debug "${tool} current: ${current_version}, latest: ${latest_version}"
 
 		if [ "${current_version}" != "${latest_version}" ]; then
-			info "update available for ${tool}"
+			info "update available for ${tool}, applying update..."
 
-			debug "applying update..."
+			debug "installing ${tool}@${latest_version}"
 			asdf install "${tool}" "${latest_version}"
+			debug "applying ${tool}@${latest_version} locally"
 			asdf local "${tool}" "${latest_version}"
 
 			remaining_capacity=$((remaining_capacity - 1))